requirements.txt
Brotli==1.0.9
build==1.0.3
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cssselect==1.1.0
cssselect2==0.6.0
cssutils==2.5.1
exceptiongroup==1.2.0
fonttools==4.34.4
html5lib==1.1
idna==3.3
iniconfig==2.0.0
lxml==4.9.1
Markdown==3.4.1
packaging==23.2
Pillow==9.2.0
pluggy==1.4.0
premailer==3.10.0
pycparser==2.21
pydyf==0.2.0
pyphen==0.14.0
pyproject_hooks==1.0.0
pytest==8.0.0
python-dotenv==0.20.0
requests==2.31.0
six==1.16.0
tinycss2==1.1.1
tomli==2.0.1
urllib3==2.2.0
weasyprint==56.1
webencodings==0.5.1
yagmail==0.15.293
zopfli==0.2.1
